Fix nasty data stream bug
Setting MSG_WAITALL in the flags argument to recv does not guarantee
that the requested number of bytes are returned. recv may, and does,
some times return early having read only part of the requested amount of
data. This was causing the TCP stream to become misaligned causing
message unpacking to fail with apparently invalid messages sizes and
attributes. The function SVR_Net_recv guarantees the full amount is
received except in the case of an error or the remote socket pair being
closed.