Fix bug in SVR_Arena_sprintf
In some cases the way that SVR_Arena_reserve maintains alignment causes
there to be a disagreement with where sprintf is writing which caused
the wrong pointer to be returned. This new implementation is robust
while sacrificing a small performance increase when space is available
in the first arena block.